+++ clickhouse/tests/test_e2e.py
@@ -3,14 +3,15 @@
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import pytest
 
+from datadog_checks.dev.utils import get_metadata_metrics
+
 from .common import CLICKHOUSE_VERSION
-from .metrics import get_metrics
+from .metrics import OPTIONAL_METRICS, get_metrics
 
 pytestmark = pytest.mark.e2e
 
 
 def test_check(dd_agent_check, instance):
-    # We do not do aggregator.assert_all_metrics_covered() because depending on timing, some other metrics may appear
     aggregator = dd_agent_check(instance, rate=True)
     server_tag = 'server:{}'.format(instance['server'])
     port_tag = 'port:{}'.format(instance['port'])
@@ -28,3 +29,9 @@ def test_check(dd_agent_check, instance):
         tags=[server_tag, port_tag, 'db:default', 'foo:bar', 'dictionary:test'],
         at_least=1,
     )
+
+    for metric in OPTIONAL_METRICS:
+        aggregator.assert_metric(metric, at_least=0)
+
+    aggregator.assert_all_metrics_covered()
+    aggregator.assert_metrics_using_metadata(get_metadata_metrics())