+++ mongo/tests/test_mongo.py
@@ -52,6 +52,7 @@ METRIC_VAL_CHECKS_OLD = {
 pytestmark = [pytest.mark.usefixtures('dd_environment'), pytest.mark.integration]
 
 
+@common.standalone
 @pytest.mark.parametrize(
     'instance_authdb',
     [
@@ -60,7 +61,7 @@ pytestmark = [pytest.mark.usefixtures('dd_environment'), pytest.mark.integration
         pytest.param(common.INSTANCE_AUTHDB_LEGACY_CONFIG, id='legacy'),
     ],
 )
-def test_mongo(aggregator, check, instance_authdb, dd_run_check):
+def test_mongo_authdb(aggregator, check, instance_authdb, dd_run_check):
     check = check(instance_authdb)
     dd_run_check(check)
 
@@ -74,11 +75,12 @@ def test_mongo(aggregator, check, instance_authdb, dd_run_check):
     aggregator.assert_metrics_using_metadata(get_metadata_metrics(), check_submission_type=True)
 
 
+@common.standalone
 @pytest.mark.parametrize(
     'instance_user',
     [pytest.param(common.INSTANCE_USER, id='standard'), pytest.param(common.INSTANCE_USER_LEGACY_CONFIG, id='legacy')],
 )
-def test_mongo2(aggregator, check, instance_user, dd_run_check):
+def test_mongo_db_test(aggregator, check, instance_user, dd_run_check):
     check = check(instance_user)
     dd_run_check(check)
 
@@ -95,6 +97,7 @@ def test_mongo2(aggregator, check, instance_user, dd_run_check):
     aggregator.assert_metrics_using_metadata(get_metadata_metrics(), check_submission_type=True)
 
 
+@common.shard
 def test_mongo_arbiter(aggregator, check, instance_arbiter, dd_run_check):
     check = check(instance_arbiter)
     dd_run_check(check)
@@ -118,7 +121,7 @@ def test_mongo_arbiter(aggregator, check, instance_arbiter, dd_run_check):
         'mongodb.replset.state': 7,
     }
     expected_tags = [
-        'server:mongodb://testUser:*****@localhost:27020/',
+        'server:mongodb://localhost:27020/',
         'replset_name:shard01',
         'replset_state:arbiter',
         'sharding_cluster_role:shardsvr',
@@ -127,6 +130,7 @@ def test_mongo_arbiter(aggregator, check, instance_arbiter, dd_run_check):
         aggregator.assert_metric(metric, value, expected_tags, count=1)
 
 
+@common.standalone
 def test_mongo_old_config(aggregator, check, instance, dd_run_check):
     check = check(instance)
     dd_run_check(check)
@@ -141,20 +145,7 @@ def test_mongo_old_config(aggregator, check, instance, dd_run_check):
     aggregator.assert_metrics_using_metadata(get_metadata_metrics(), check_submission_type=True)
 
 
-def test_mongo_old_config_2(aggregator, check, instance, dd_run_check):
-    check = check(instance)
-    dd_run_check(check)
-
-    metric_names = aggregator.metric_names
-    assert metric_names
-
-    for metric_name in metric_names:
-        if metric_name in METRIC_VAL_CHECKS_OLD:
-            metric = aggregator.metrics(metric_name)[0]
-            assert METRIC_VAL_CHECKS_OLD[metric_name](metric.value)
-    aggregator.assert_metrics_using_metadata(get_metadata_metrics(), check_submission_type=True)
-
-
+@common.standalone
 def test_mongo_1valid_and_1invalid_custom_queries(
     aggregator, check, instance_1valid_and_1invalid_custom_queries, dd_run_check
 ):
@@ -167,6 +158,7 @@ def test_mongo_1valid_and_1invalid_custom_queries(
     aggregator.assert_metric("dd.custom.mongo.query_a.amount", count=0)
 
 
+@common.standalone
 def test_mongo_custom_queries(aggregator, check, instance_custom_queries, dd_run_check):
     # Run the check against our running server
     check = check(instance_custom_queries)
@@ -204,6 +196,7 @@ def test_mongo_custom_queries(aggregator, check, instance_custom_queries, dd_run
     aggregator.assert_metric('dd.mongodb.custom.queries_slower_than_60sec.secs_running', metric_type=aggregator.GAUGE)
 
 
+@common.standalone
 def test_mongo_custom_query_with_empty_result_set(aggregator, check, instance_user, caplog, dd_run_check):
     instance_user['custom_queries'] = [
         {
@@ -229,6 +222,7 @@ def test_mongo_custom_query_with_empty_result_set(aggregator, check, instance_us
     aggregator.assert_metric('dd.custom.mongo.query_a.amount', count=0)
 
 
+@common.shard
 def test_mongo_replset(instance_shard, aggregator, check, dd_run_check):
     mongo_check = check(instance_shard)
     dd_run_check(mongo_check)
@@ -256,6 +250,7 @@ def test_mongo_replset(instance_shard, aggregator, check, dd_run_check):
     aggregator.assert_metrics_using_metadata(get_metadata_metrics(), check_submission_type=True)
 
 
+@common.standalone
 def test_metadata(check, instance, datadog_agent):
     check = check(instance)
     check.check_id = 'test:123'
@@ -267,6 +262,7 @@ def test_metadata(check, instance, datadog_agent):
     datadog_agent.assert_metadata_count(len(version_metadata) + 2)
 
 
+@common.shard
 def test_refresh_role(instance_shard, aggregator, check, dd_run_check):
     mongo_check = check(instance_shard)
     dd_run_check(mongo_check)