+++ mysql/CHANGELOG.md
@@ -1,6 +1,12 @@
 # CHANGELOG - mysql
 
-## 8.4.1 / 2022-07-08
+## 8.5.0 / 2022-08-05
+
+* [Added] Add MySQL user connections metric. See [#12657](https://github.com/DataDog/integrations-core/pull/12657).
+* [Fixed] Dependency updates. See [#12653](https://github.com/DataDog/integrations-core/pull/12653).
+* [Fixed] Pin `pymysql` to `0.10.1`. See [#12612](https://github.com/DataDog/integrations-core/pull/12612).
+
+## 8.4.1 / 2022-07-08 / Agent 7.38.0
 
 * [Fixed] Fix bug where mysql table row stats were not being collected. See [#12472](https://github.com/DataDog/integrations-core/pull/12472).
 
@@ -27,7 +33,7 @@
 * [Added] Add query_truncated field on activity rows. See [#11886](https://github.com/DataDog/integrations-core/pull/11886).
 * [Fixed] Fix uncommented parent options. See [#12013](https://github.com/DataDog/integrations-core/pull/12013).
 
-## 8.2.3 / 2022-05-26
+## 8.2.3 / 2022-05-26 / Agent 7.36.1
 
 * [Fixed] Revert mysql.net.connections metric type. See [#12088](https://github.com/DataDog/integrations-core/pull/12088).