+++ windows_performance_counters/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG - Windows performance counters
 
+## 1.3.0 / 2022-08-05
+
+* [Added] Add functionality to load the legacy version of the integration. See [#12396](https://github.com/DataDog/integrations-core/pull/12396).
+
 ## 1.2.0 / 2022-04-05 / Agent 7.36.0
 
 * [Added] Add metric_patterns options to filter all metric submission by a list of regexes. See [#11695](https://github.com/DataDog/integrations-core/pull/11695).